requirements.txt
deemix
deezer-py
python-magic
